change: move services/routes.tsx to configs/routes.tsx
