+++ src/main.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import {RouterProvider} from 'react-router-dom'
 
-import { router } from "@/services/routes";
+import { router } from "@/configs/routes";
 
 import 'bootstrap-icons/font/bootstrap-icons.min.css'
 import 'sweetalert2/dist/sweetalert2.min.css'
